requirements.txt
Flask
web3
py-solc-x
